move styled select as requested
