+++ shared-ui/style/typography.ts
@@ -66,23 +66,4 @@ const StyledLink = styled.a`
   }
 `;
 
-const StyledDropdown = styled.select`
-  border-color: ${colors.WHITE};
-  background-color: ${colors.BUTTON_DARK_GREEN};
-  color: ${colors.WHITE};
-  font-family: ${fonts.nunitoSansSemibold};
-  border-radius: 0.3em;
-  height: 3em;
-  padding-left: 0.5em;
-  padding-right: 0.5em;
-
-  @media ${min.mobile} {
-    font-size: 1.2em;
-  }
-  @media ${min.tablet} {
-    font-size: 1.25em;
-  }
-`;
-
-
-export { H1, H2, H3, H4, P, fonts, StyledLink, StyledDropdown };
+export { H1, H2, H3, H4, P, fonts, StyledLink };
